requirements.txt
qq-botpy >= 1.1.1
requests == 2.27.1
urllib3 == 1.26.9
configparser == 5.2.0